reorganize api route handling
- move server init steps into index.js
- move route setup into routes.js
- add proposal and slate controllers
- define routes/controllers separately
- move proposalSchema to utils
